MATE-150 : [FEAT] 굿즈거래 채팅 엔티티를 MongoDB 도큐먼트로 변경
